Rename parser.js to vast_parser.js
