+++ src/parser/vast_parser.js
